requirements.txt
flask
python-dotenv
pymongo
bcrypt
flask-mail
flask-redis
Flask-Session